Extract builder methods from `PartialGlobalVertex`
