+++ crates/fj-kernel/src/partial/objects/vertex.rs
@@ -1,6 +1,7 @@
 use fj_math::Point;
 
 use crate::{
+    builder::GlobalVertexBuilder,
     objects::{Curve, GlobalVertex, Objects, Surface, SurfaceVertex, Vertex},
     partial::{HasPartial, MaybePartial},
     storage::Handle,
@@ -241,34 +242,6 @@ impl PartialGlobalVertex {
             self.position = Some(position.into());
         }
         self
-    }
-
-    /// Update partial global vertex from the given curve and position on it
-    pub fn from_curve_and_position(
-        self,
-        curve: impl Into<MaybePartial<Curve>>,
-        position: impl Into<Point<1>>,
-    ) -> Self {
-        let curve = curve.into().into_partial();
-
-        let path = curve.path().expect(
-            "Need path to create `GlobalVertex` from curve and position",
-        );
-        let surface = curve.surface().expect(
-            "Need surface to create `GlobalVertex` from curve and position",
-        );
-
-        let position_surface = path.point_from_path_coords(position);
-        self.from_surface_and_position(&surface, position_surface)
-    }
-
-    /// Update partial global vertex from the given surface and position on it
-    pub fn from_surface_and_position(
-        self,
-        surface: &Surface,
-        position: impl Into<Point<2>>,
-    ) -> Self {
-        self.with_position(Some(surface.point_from_surface_coords(position)))
     }
 
     /// Build a full [`GlobalVertex`] from the partial global vertex
